# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Get Am Nice API — Salone cultural companion app
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface UserProfile {
  id: number;
  clerkUserId: string;
  displayName?: string | null;
  email?: string | null;
  createdAt: string;
}

export interface UpdateProfileBody {
  displayName: string;
}

export interface DictionaryLookupBody {
  /** The Krio or Salone word/phrase to look up */
  term: string;
}

export interface DictionaryEntry {
  term: string;
  definition: string;
  culturalContext: string;
  usageExamples: string[];
  pronunciation?: string | null;
  partOfSpeech?: string | null;
}

export interface LexiconEntry {
  id: number;
  term: string;
  definition: string;
  culturalContext: string;
  usageExamples: string[];
  pronunciation?: string | null;
  partOfSpeech?: string | null;
  /** User's own notes or corrections */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateLexiconEntryBody {
  term: string;
  definition: string;
  culturalContext: string;
  usageExamples: string[];
  pronunciation?: string | null;
  partOfSpeech?: string | null;
  notes?: string | null;
}

export interface UpdateLexiconEntryBody {
  definition?: string;
  culturalContext?: string;
  usageExamples?: string[];
  pronunciation?: string | null;
  partOfSpeech?: string | null;
  notes?: string | null;
}

export interface Event {
  id: number;
  title: string;
  description?: string | null;
  location: string;
  city?: string | null;
  country?: string | null;
  eventDate: string;
  venue?: string | null;
  ticketUrl?: string | null;
  submittedBy?: string | null;
  approved: boolean;
  createdAt: string;
}

export interface SubmitEventBody {
  title: string;
  description?: string | null;
  location: string;
  city?: string | null;
  country?: string | null;
  eventDate: string;
  venue?: string | null;
  ticketUrl?: string | null;
  submittedBy?: string | null;
}

export type ListEventsParams = {
  /**
   * Filter events by city or country
   */
  location?: string;
};