# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Get Am Nice API — Salone cultural companion app
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get current user profile
 */
export const GetProfileResponse = zod.object({
  id: zod.number(),
  clerkUserId: zod.string(),
  displayName: zod.string().nullish(),
  email: zod.string().nullish(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Update user profile display name
 */
export const UpdateProfileBody = zod.object({
  displayName: zod.string(),
});

export const UpdateProfileResponse = zod.object({
  id: zod.number(),
  clerkUserId: zod.string(),
  displayName: zod.string().nullish(),
  email: zod.string().nullish(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Look up a Krio or Salone word/phrase using Claude AI
 */
export const DictionaryLookupBody = zod.object({
  term: zod.string().describe("The Krio or Salone word\/phrase to look up"),
});

export const DictionaryLookupResponse = zod.object({
  term: zod.string(),
  definition: zod.string(),
  culturalContext: zod.string(),
  usageExamples: zod.array(zod.string()),
  pronunciation: zod.string().nullish(),
  partOfSpeech: zod.string().nullish(),
});

/**
 * @summary Get today's featured Salone term
 */
export const GetTermOfTheDayResponse = zod.object({
  term: zod.string(),
  definition: zod.string(),
  culturalContext: zod.string(),
  usageExamples: zod.array(zod.string()),
  pronunciation: zod.string().nullish(),
  partOfSpeech: zod.string().nullish(),
});

/**
 * @summary List saved lexicon entries for the authenticated user
 */
export const ListLexiconResponseItem = zod.object({
  id: zod.number(),
  term: zod.string(),
  definition: zod.string(),
  culturalContext: zod.string(),
  usageExamples: zod.array(zod.string()),
  pronunciation: zod.string().nullish(),
  partOfSpeech: zod.string().nullish(),
  notes: zod.string().nullish().describe("User's own notes or corrections"),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const ListLexiconResponse = zod.array(ListLexiconResponseItem);

/**
 * @summary Save a new entry to the user's personal lexicon
 */
export const CreateLexiconEntryBody = zod.object({
  term: zod.string(),
  definition: zod.string(),
  culturalContext: zod.string(),
  usageExamples: zod.array(zod.string()),
  pronunciation: zod.string().nullish(),
  partOfSpeech: zod.string().nullish(),
  notes: zod.string().nullish(),
});

/**
 * @summary Edit a saved lexicon entry (e.g. to correct AI errors)
 */
export const UpdateLexiconEntryParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateLexiconEntryBody = zod.object({
  definition: zod.string().optional(),
  culturalContext: zod.string().optional(),
  usageExamples: zod.array(zod.string()).optional(),
  pronunciation: zod.string().nullish(),
  partOfSpeech: zod.string().nullish(),
  notes: zod.string().nullish(),
});

export const UpdateLexiconEntryResponse = zod.object({
  id: zod.number(),
  term: zod.string(),
  definition: zod.string(),
  culturalContext: zod.string(),
  usageExamples: zod.array(zod.string()),
  pronunciation: zod.string().nullish(),
  partOfSpeech: zod.string().nullish(),
  notes: zod.string().nullish().describe("User's own notes or corrections"),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Remove an entry from the user's lexicon
 */
export const DeleteLexiconEntryParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List upcoming Salone music events, optionally filtered by location
 */
export const ListEventsQueryParams = zod.object({
  location: zod.coerce
    .string()
    .optional()
    .describe("Filter events by city or country"),
});

export const ListEventsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string().nullish(),
  location: zod.string(),
  city: zod.string().nullish(),
  country: zod.string().nullish(),
  eventDate: zod.coerce.date(),
  venue: zod.string().nullish(),
  ticketUrl: zod.string().nullish(),
  submittedBy: zod.string().nullish(),
  approved: zod.boolean(),
  createdAt: zod.coerce.date(),
});
export const ListEventsResponse = zod.array(ListEventsResponseItem);

/**
 * @summary Submit a new event for listing
 */
export const SubmitEventBody = zod.object({
  title: zod.string(),
  description: zod.string().nullish(),
  location: zod.string(),
  city: zod.string().nullish(),
  country: zod.string().nullish(),
  eventDate: zod.coerce.date(),
  venue: zod.string().nullish(),
  ticketUrl: zod.string().nullish(),
  submittedBy: zod.string().nullish(),
});

/**
 * @summary Get a short preview of the next 3 upcoming events (for the home screen)
 */
export const GetUpcomingEventsPreviewResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string().nullish(),
  location: zod.string(),
  city: zod.string().nullish(),
  country: zod.string().nullish(),
  eventDate: zod.coerce.date(),
  venue: zod.string().nullish(),
  ticketUrl: zod.string().nullish(),
  submittedBy: zod.string().nullish(),
  approved: zod.boolean(),
  createdAt: zod.coerce.date(),
});
export const GetUpcomingEventsPreviewResponse = zod.array(
  GetUpcomingEventsPreviewResponseItem,
);