# lib/api-zod/src/generated/types/createLexiconEntryBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Get Am Nice API — Salone cultural companion app
 * OpenAPI spec version: 0.1.0
 */

export interface CreateLexiconEntryBody {
  term: string;
  definition: string;
  culturalContext: string;
  usageExamples: string[];
  pronunciation?: string | null;
  partOfSpeech?: string | null;
  notes?: string | null;
}