# lib/api-zod/src/generated/types/dictionaryLookupBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Get Am Nice API — Salone cultural companion app
 * OpenAPI spec version: 0.1.0
 */

export interface DictionaryLookupBody {
  /** The Krio or Salone word/phrase to look up */
  term: string;
}