# lib/api-zod/src/generated/types/event.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Get Am Nice API — Salone cultural companion app
 * OpenAPI spec version: 0.1.0
 */

export interface Event {
  id: number;
  title: string;
  description?: string | null;
  location: string;
  city?: string | null;
  country?: string | null;
  eventDate: Date;
  venue?: string | null;
  ticketUrl?: string | null;
  submittedBy?: string | null;
  approved: boolean;
  createdAt: Date;
}