# lib/api-zod/src/generated/types/lexiconEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Get Am Nice API — Salone cultural companion app
 * OpenAPI spec version: 0.1.0
 */

export interface LexiconEntry {
  id: number;
  term: string;
  definition: string;
  culturalContext: string;
  usageExamples: string[];
  pronunciation?: string | null;
  partOfSpeech?: string | null;
  /** User's own notes or corrections */
  notes?: string | null;
  createdAt: Date;
  updatedAt: Date;
}