# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Get Am Nice API — Salone cultural companion app
 * OpenAPI spec version: 0.1.0
 */

export interface UserProfile {
  id: number;
  clerkUserId: string;
  displayName?: string | null;
  email?: string | null;
  createdAt: Date;
}